% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attribute_filter.R
\name{attribute_filter}
\alias{attribute_filter}
\title{Return subgraph where attributes match conditions}
\usage{
attribute_filter(graph, attr_expression)
}
\arguments{
\item{graph}{Graph of class '\code{igraph}'}

\item{attr_expression}{A logical expression defined in terms of the attribute names of \code{graph}.
only vertices where the expression evaluates to \code{TRUE} are kept.}
}
\description{
Use `attribute_filter()` to find cases where conditions are true given vertex attributes.
}
