% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attribute_seed.R
\name{attribute_seed}
\alias{attribute_seed}
\title{Return seeded graph.}
\usage{
attribute_seed(graph, attr_expression)
}
\arguments{
\item{graph}{Graph of class '\code{igraph}'}

\item{attr_expression}{A logical expression defined in terms of the attribute names of \code{graph}.
only vertices where the expression evaluates to \code{TRUE} are kept.}
}
\description{
Use `attribute_seed()` to seed network nodes with diffusion points
based on given vertex attributes.
}
