% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_diffusion.R
\name{calc_diffusion}
\alias{calc_diffusion}
\title{Network propagation}
\usage{
calc_diffusion(graph, method = "raw")
}
\arguments{
\item{graph}{Graph of class '\code{igraph}'. Must include node
attributes '\code{name}' and '\code{seed}'.}

\item{method}{String. Kernel for diffusion. Passed to \code{diffusion()}
from the \code{diffuStats} package.}
}
\value{
Graph of class '\code{igraph}' with new attribute, '\code{diffusion_score}'.
}
\description{
Calculate various measures of network propagation
}
