% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_random_walk.R
\name{calc_random_walk}
\alias{calc_random_walk}
\title{Calculate network propagation by random walk}
\usage{
calc_random_walk(graph, restart_value = 0.75, norm_method = "column")
}
\arguments{
\item{graph}{Graph of class '\code{igraph}'. Must include node
attributes '\code{name}' and '\code{seed}'.}

\item{method}{String. Method for random walk. Passed to \code{dRWR()}
from the \code{dnet} package.}
}
\value{
Graph of class '\code{igraph}' with new attribute, '\code{propagation_score}'.
}
\description{
Calculate various measures of random walk
}
