% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_create.R
\name{db_create}
\alias{db_create}
\title{db_create}
\usage{
db_create(vers = "11", spec = 9606, thresh = 950)
}
\arguments{
\item{vers}{stingDB version}

\item{spec}{species code #9606 for human}

\item{thresh}{score_threshold}
}
\description{
add two numbers
}
