% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_to_vert_attr.R
\name{df_to_vert_attr}
\alias{df_to_vert_attr}
\title{Set the vertex attributes of a graph}
\usage{
df_to_vert_attr(graph, df, common, attr_name, new_attr_name = NULL)
}
\arguments{
\item{graph}{Graph of class '\code{igraph}'}

\item{df}{A data frame that has a column that contains vertex IDs
and at least one column that has data that will become a vertex attribute of \code{graph}.}

\item{common}{A character string providing the column name that contains
the vertex IDs (i.e., the 'name' vertex attribute of the graph).}

\item{attr_name}{A character vector providing the name(s) of the
column of \code{df} that will be added as vertex attributes to \code{graph}}

\item{new_attr_name}{A character vector providing the name(s) of the
new vertex attribute(s). If \code{NULL} then the name of the
new attribute(s) will match \code{attr_name}.}
}
\description{
Given a graph and a data frame where vertex
names correspond to a column in the data frame, map data
in the data frame to attributes of the vertices in the graph.
}
