% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param_heatmap.R
\name{param_heatmap}
\alias{param_heatmap}
\title{Parameter Grid Heatmap}
\usage{
param_heatmap(results, plot_margin = c(8, 5))
}
\arguments{
\item{results}{Object of class '\code{dataframe}'. Parameter grid results.}

\item{plot_margin}{Object of class '\code{list}'. Allows adjustment of plot margins.}
}
\value{
Object of class '\code{image}' showing z-score statistics derived from 
parameter_grid results
}
\description{
Create a heatmap of parameter grid results.
}
