% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_graph.R
\name{plot_graph}
\alias{plot_graph}
\title{Plot final network.}
\usage{
plot_graph(graph, method, gene_list)
}
\arguments{
\item{graph}{Graph of class '\code{igraph}'. Must include node
attributes '\code{name}' and '\code{seed}'.}

\item{method}{String. Name of node attribute that should correspond to node size.}

\item{gene_list}{List of strings. Symbols of primary genes of interest.}
}
\description{
Plot network with standard visualization specifications.
}
