% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/propagation_pipeline.R
\name{propagation_pipeline}
\alias{propagation_pipeline}
\title{Pipeline function to run complete noderank network propagation workflow.}
\usage{
propagation_pipeline(
  deg,
  ppi = NULL,
  string_db = NULL,
  edge_conf_score_min,
  logFC_min,
  pvalue_max,
  method = "raw",
  min_diff_score = 0.15,
  causal_gene_symbol,
  export_network = FALSE,
  export_dir = NULL,
  sim_method = "jaccard",
  n_sim = 9999,
  weighted = FALSE,
  ...
)
}
\arguments{
\item{deg}{Object of class '\code{dataframe}'. Differential gene expression analysis
results including log fold-changes and p-values,
i.e. the output of \code{topTable()} from limma or \code{results()} from DESeq2.}

\item{edge_conf_score_min}{Numeric. A The minimum confidence score between edges in the
STRING protein-protein interaction network.}

\item{logFC_min}{Numeric. The minimum log2 fold-change value for a gene
to be included in the final network.}

\item{pvalue_max}{Numeric. The maximum p-value value for a gene
to be included in the final network.}

\item{method}{String. The method passed to the  diffusion() function
from the diffuStats() package. One of the following methods:
\itemize{
\item raw
\item ml
\item gm
\item ber_s
\item mc
\item ber_p
\item z
}}

\item{min_diff_score}{Numeric. The diffusion score threshold for filtering
the final network.}

\item{causal_gene_symbol}{String. The gene symbol associated with the
causal gene.}

\item{export_network}{Logical. If TRUE, the network object will be returned.}

\item{export_dir}{String. If export_network = TRUE, the network object will be
saved using the file path provided in export_dir.}

\item{sim_method}{String. The method for calculating the similarity between
each gene in the final network and the causal gene. One of the following:
#' \itemize{
\item jaccard
\item dice
\item invlogweighted }}

\item{n_sim}{Numeric. The number of simulations passed to \code{evaluate_performance()}.}
}
\value{
A list of length 4:
\describe{
  \item{network}{Object of class '\code{igraph}'. The network of important genes.}
  \item{top_genes}{An annotated dataframe of genes ranked by diffusion score.}
  \item{mean_score}{The mean score of the network, which is equal to the average
  similarity score between each node in the network and the causal gene.}
  \item{pvalue}{The p-value associated with the \code{mean_score}.}
}
}
\description{
Given a dataframe of differential gene expression results and a set
of parameters, '\code{centrality_pipeline()}' will generate network of the most
important genes, ranked by the specified centrality measure.
}
