% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_graph.R
\name{subset_graph}
\alias{subset_graph}
\title{Subset the loaded igraph based on edge confidence}
\usage{
subset_graph(graph, score_threshold)
}
\arguments{
\item{graph}{Graph of class '\code{igraph}'}

\item{score_threshold}{Integer value of the edge confidence'}
}
\description{
Given an iGraph object from StringDb, reduce the
full graph to a subset based on edge confidence scores. This
function will improve the performance of multiple calls to
the iGraph object for the parameter optimization.
}
