% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attribute_seed.R
\name{attribute_seed}
\alias{attribute_seed}
\title{Return seeded graph.}
\usage{
attribute_seed(graph, attr_expression)
}
\arguments{
\item{graph}{Graph of class '\code{igraph}'}

\item{attr_expression}{A logical expression defined in terms of the attribute names of \code{graph}.
only vertices where the expression evaluates to \code{TRUE} are kept.}
}
\description{
Use `attribute_seed()` to seed network nodes with diffusion points
based on given vertex attributes.
}
\examples{
# generate a graph
vertex_meta <- data.frame('name' = c('a','b','c','d'), value = c(1, 5, 10, 3))
edge_list <- data.frame('v1' = c('a', 'a', 'd', 'd'), 'v2' = c('b', 'c', 'a', 'c'))
graph <- igraph::graph_from_data_frame(d = edge_list, vertices = vertex_meta, directed = F)

# examine initial graph
igraph::vertex_attr(graph)

# seed graph by node value
seeded_graph <- attribute_seed(graph = graph, value > 4)

# examine seeded graph
igraph::vertex_attr(seeded_graph)

# there is now a node attribute 'seed', which is equal to 1 for nodes b and c,
# and 0 for nodes a and d.
}
