% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_create.R
\name{db_create}
\alias{db_create}
\title{db_create}
\usage{
db_create(vers = "11", spec = 9606, thresh = 950)
}
\arguments{
\item{vers}{A string representing the stringDB version.}

\item{spec}{A number specifying the species code (default is 9606 for humans.)}

\item{thresh}{A number specifying score_threshold (a measure of interaction strength.)}
}
\value{
An iGraph object representing the ppi network.
}
\description{
Download protein-protein interaction (ppi) network and return in the form of
an iGraph object for processing.
}
